+++ LegendWeathers/Config.cs
@@ -21,6 +21,7 @@ namespace LegendWeathers
         public readonly ConfigEntry<string> bloodMoonTexture;
         public readonly ConfigEntry<float> bloodMoonSizeFactor;
         public readonly ConfigEntry<int> bloodMoonEffectsAbundance;
+        public readonly ConfigEntry<float> bloodMoonEffectsVolume;
         public readonly ConfigEntry<float> bloodMoonIntroMusicVolume;
         public readonly ConfigEntry<float> bloodMoonAmbienceMusicVolume;
         public readonly ConfigEntry<string> bloodMoonAmbienceMusicType;
@@ -46,6 +47,7 @@ namespace LegendWeathers
             bloodMoonTexture = cfg.Bind("Blood Moon", "Moon texture", "Classic", new ConfigDescription("Choose the texture used for the Blood Moon material.", new AcceptableValueList<string>("Classic", "Bright")));
             bloodMoonSizeFactor = cfg.Bind("Blood Moon", "Moon size factor", 5.5f, new ConfigDescription("The size factor of the Blood Moon material compared to the size of the vanilla sun.", new AcceptableValueRange<float>(1f, 10f)));
             bloodMoonEffectsAbundance = cfg.Bind("Blood Moon", "Terrain effects abundance", 30, new ConfigDescription("The abundance of terrain effects during the Blood Moon weather.\nHigher values will spawn more effects but may impact performance.", new AcceptableValueRange<int>(0, 100)));
+            bloodMoonEffectsVolume = cfg.Bind("Blood Moon", "Terrain effects volume", 1f, new ConfigDescription("The Blood Moon's terrain effects all have small sfx coming from them. Those sfx volumes are not very high but if you really want to reduce it then this config is for you.", new AcceptableValueRange<float>(0f, 1f)));
             bloodMoonIntroMusicVolume = cfg.Bind("Blood Moon", "Intro music volume", 0.8f, new ConfigDescription("An introduction music is played when the Blood Moon is spawned on the planet, this will last 30s.\nYou can customize the music volume here.", new AcceptableValueRange<float>(0f, 1f)));
             bloodMoonAmbienceMusicVolume = cfg.Bind("Blood Moon", "Ambience music volume", 1f, new ConfigDescription("The volume for the Blood Moon ambient music when you are outside (similar to the vanilla Eclipsed ambient music).", new AcceptableValueRange<float>(0f, 1f)));
             bloodMoonAmbienceMusicType = cfg.Bind("Blood Moon", "Ambience music type", "Blood Moon", new ConfigDescription("If you prefer to have the vanilla ambient Eclipsed music instead of the Blood Moon ambient custom music, then you can change it here.\nPlease note that the 'Ambient music volume' config will not apply if you select the Eclipsed music.", new AcceptableValueList<string>("Blood Moon", "Eclipsed")));