+++ zigzag.legendweathers/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 2.0.4
 - **Blood Moon**
+    - Added `Terrain effects volume` config, this can be used to customize the volume of the weather's terrain effects
     - Added `Ambience music volume` config, this can be used to customize the volume of the weather's ambient music
-    - Added `Ambience music type` config, you can use this to replace the weather's ambient music by the vanilla Eclipsed music
+    - Added `Ambience music type` config, you can use this to replace the weather's ambient music by the vanilla Eclipsed music if wanted
 - **Majora Moon**
     - f